Remove default argument to remove dependency to third party libraries
