+++ tests/OTAnalytics/plugin_parser/test_otvision_parser.py
@@ -35,6 +35,9 @@ from OTAnalytics.plugin_datastore.python_track_store import (
     PythonTrack,
     PythonTrackDataset,
 )
+from OTAnalytics.plugin_datastore.track_geometry_store.pygeos_store import (
+    PygeosTrackGeometryDataset,
+)
 from OTAnalytics.plugin_parser import dataformat_versions, ottrk_dataformat
 from OTAnalytics.plugin_parser.json_parser import write_json, write_json_bz2
 from OTAnalytics.plugin_parser.otvision_parser import (
@@ -212,6 +215,7 @@ class TestOttrkParser:
         detection_parser = PythonDetectionParser(
             calculator,
             mocked_track_repository,
+            PygeosTrackGeometryDataset.from_track_dataset,
             track_length_limit=DEFAULT_TRACK_LENGTH_LIMIT,
         )
         return OttrkParser(detection_parser)
@@ -247,7 +251,11 @@ class TestOttrkParser:
             ["person", "bus", "boat", "truck", "car", "motorcycle", "bicycle", "train"]
         )
         assert_track_datasets_equal(
-            parse_result.tracks, PythonTrackDataset.from_list([expected_track])
+            parse_result.tracks,
+            PythonTrackDataset.from_list(
+                [expected_track],
+                PygeosTrackGeometryDataset.from_track_dataset,
+            ),
         )
         assert (
             parse_result.detection_metadata.detection_classes
@@ -285,6 +293,7 @@ class TestPythonDetectionParser:
         return PythonDetectionParser(
             mocked_classificator,
             mocked_track_repository,
+            PygeosTrackGeometryDataset.from_track_dataset,
         )
 
     def test_parse_detections_output_has_same_order_as_input(
@@ -337,7 +346,8 @@ class TestPythonDetectionParser:
         result_unsorted_input = parser.parse_tracks(unsorted_detections, metadata_video)
 
         expected_sorted = PythonTrackDataset.from_list(
-            [track_builder_setup_with_sample_data.build_track()]
+            [track_builder_setup_with_sample_data.build_track()],
+            PygeosTrackGeometryDataset.from_track_dataset,
         )
         assert_track_datasets_equal(result_sorted_input, expected_sorted)
         assert_track_datasets_equal(result_unsorted_input, expected_sorted)
@@ -375,7 +385,9 @@ class TestPythonDetectionParser:
 
         result_sorted_input = parser.parse_tracks(detections, metadata_video)
 
-        expected_sorted = PythonTrackDataset.from_list([merged_track])
+        expected_sorted = PythonTrackDataset.from_list(
+            [merged_track], PygeosTrackGeometryDataset.from_track_dataset
+        )
 
         assert_track_datasets_equal(result_sorted_input, expected_sorted)
         mocked_classificator.calculate.assert_called_once_with(all_detections)
@@ -396,6 +408,7 @@ class TestPythonDetectionParser:
         parser = PythonDetectionParser(
             ByMaxConfidence(),
             mocked_track_repository,
+            PygeosTrackGeometryDataset.from_track_dataset,
             track_length_limit,
         )
         detections: list[dict] = (
